requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
asyncpg==0.30.0
requests==2.32.3
feedparser==6.0.11
jinja2==3.1.6
python-multipart==0.0.20
aiofiles==24.1.0
psycopg2-binary==2.9.10
pydantic==2.11.5
